backend/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.6.2.post1 ; python_version >= "3.10" and python_version < "4.0"
appdirs==1.4.4 ; python_version >= "3.10" and python_version < "4.0"
base58==2.1.1 ; python_version >= "3.10" and python_version < "4"
certifi==2024.8.30 ; python_version >= "3.10" and python_version < "4"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
cytoolz==1.0.0 ; python_version >= "3.10" and python_version < "4" and implementation_name == "cpython"
dnspython==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
ecdsa==0.19.0 ; python_version >= "3.10" and python_version < "4"
eth-hash==0.7.0 ; python_version >= "3.10" and python_version < "4"
eth-keys==0.6.0 ; python_version >= "3.10" and python_version < "4"
eth-typing==5.0.1 ; python_version >= "3.10" and python_version < "4"
eth-utils==5.1.0 ; python_version >= "3.10" and python_version < "4"
exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.115.4 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.54.1 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.6 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
ipfs-toolkit==0.5.26 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.7 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.9.2 ; python_version >= "3.10" and python_version < "4.0"
more-itertools==10.5.0 ; python_version >= "3.10" and python_version < "4"
multiaddr==0.0.9 ; python_version >= "3.10" and python_version < "4.0"
netaddr==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.1.2 ; python_version < "4.0" and python_version >= "3.10"
opencv-python==4.10.0.84 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.10" and python_version < "4.0"
py-bip39-bindings==0.1.11 ; python_version >= "3.10" and python_version < "4"
py-ed25519-zebra-bindings==1.1.0 ; python_version >= "3.10" and python_version < "4"
py-sr25519-bindings==0.2.0 ; python_version >= "3.10" and python_version < "4"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4"
pycryptodome==3.21.0 ; python_version >= "3.10" and python_version < "4"
pydantic-core==2.23.4 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.10" and python_version < "4.0"
pymongo==4.10.1 ; python_version >= "3.10" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.10" and python_version < "4"
pyparsing==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
robonomics-interface==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
scalecodec==1.2.11 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.41.2 ; python_version >= "3.10" and python_version < "4.0"
substrate-interface==1.7.11 ; python_version >= "3.10" and python_version < "4"
termcolor==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
toolz==1.0.0 ; python_version >= "3.10" and python_version < "4" and implementation_name == "pypy" or python_version >= "3.10" and python_version < "4" and implementation_name == "cpython"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.32.0 ; python_version >= "3.10" and python_version < "4.0"
varint==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
websocket-client==1.8.0 ; python_version >= "3.10" and python_version < "4"
xxhash==3.5.0 ; python_version >= "3.10" and python_version < "4"
attrs==24.2.0
brother-ql==0.9.4
click==8.1.7
future==1.0.0
packbits==0.6
Pillow==9.5.0
pyusb==1.2.1
qrcode==8.0
